20.06 - 14.26
zrobione zarządzanie tagami przez admina

+++ app/src/Controller/TagController.php
@@ -38,19 +38,21 @@ class TagController extends AbstractController
      */
     public function index(Request $request, TagRepository $repository, PaginatorInterface $paginator): Response
     {
-//        $pagination = $paginator->paginate(
-////            $repository->queryAll(),
-////            $request->query->getInt('page', 1),
-////            Tag::NUMBER_OF_ITEMS
-////        );
-////
-////        return $this->render(
-////            'tag/index.html.twig',
-////            ['pagination' => $pagination]
-////        );
+        if ($this->isGranted('ROLE_ADMIN')==false){
+            $this->addFlash('warning', 'message.not_have_access');
+
+            return $this->redirectToRoute('main_index');
+        }
+
+        $pagination = $paginator->paginate(
+            $repository->queryAll(),
+            $request->query->getInt('page', 1),
+            Tag::NUMBER_OF_ITEMS
+        );
+
         return $this->render(
-            'ingredient/index.html.twig',
-            ['ingredients' => $repository->findAll()]
+            'tag/index.html.twig',
+            ['tags' => $pagination]
         );
     }
 
@@ -80,46 +82,10 @@ class TagController extends AbstractController
             [
                 'pagination' => $pagination,
                 'tag' => $tag,
-//                'videos' =>$tagRepository -> findRecipeWithThisTag($id)
             ]
         );
     }
-    /**
-     * New action.
-     *
-     * @param \Symfony\Component\HttpFoundation\Request $request    HTTP request
-     * @param \App\Repository\TagRepository        $repository Tag repository
-     *
-     * @return \Symfony\Component\HttpFoundation\Response HTTP response
-     *
-     * @throws \Doctrine\ORM\ORMException
-     * @throws \Doctrine\ORM\OptimisticLockException
-     *
-     * @Route(
-     *     "/new",
-     *     methods={"GET", "POST"},
-     *     name="tag_new",
-     * )
-     */
-    public function new(Request $request, TagRepository $repository): Response
-    {
-        $tag = new Tag();
-        $form = $this->createForm(TagType::class, $tag);
-        $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $repository->save($tag);
-
-            $this->addFlash('success', 'message.created_successfully');
-
-            return $this->redirectToRoute('tag_index');
-        }
-
-        return $this->render(
-            'tag/new.html.twig',
-            ['form' => $form->createView()]
-        );
-    }
     /**
      * Edit action.
      *
@@ -141,6 +107,12 @@ class TagController extends AbstractController
      */
     public function edit(Request $request, Tag $tag, TagRepository $repository): Response
     {
+        if ($this->isGranted('ROLE_ADMIN')==false){
+            $this->addFlash('warning', 'message.not_have_access');
+
+            return $this->redirectToRoute('main_index');
+        }
+
         $form = $this->createForm(TagType::class, $tag, ['method' => 'PUT']);
         $form->handleRequest($request);
 
@@ -153,7 +125,7 @@ class TagController extends AbstractController
         }
 
         return $this->render(
-            'tag/editData.html.twig',
+            'tag/edit.html.twig',
             [
                 'form' => $form->createView(),
                 'tag' => $tag,
@@ -181,6 +153,12 @@ class TagController extends AbstractController
      */
     public function delete(Request $request, Tag $tag, TagRepository $repository): Response
     {
+        if ($this->isGranted('ROLE_ADMIN')==false){
+            $this->addFlash('warning', 'message.not_have_access');
+
+            return $this->redirectToRoute('main_index');
+        }
+
         $form = $this->createForm(FormType::class, $tag, ['method' => 'DELETE']);
         $form->handleRequest($request);
